requirements.txt
langgraph>=0.2.0
langchain-anthropic>=0.3.0
anthropic>=0.40.0
yfinance>=0.2.40
finnhub-python>=2.4.19
praw>=7.7.1
pandas-ta>=0.3.14b
pydantic>=2.5.0
watchdog>=4.0.0
python-dotenv>=1.0.0
pandas>=2.1.0
numpy>=1.26.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
